+++ test/python/_mockutils.py
@@ -22,7 +22,8 @@ import time
 from qiskit.result import Result
 from qiskit.backends import BaseBackend
 from qiskit.backends import BaseJob
-from qiskit.backends.models import BackendProperties
+from qiskit.backends.models import BackendProperties, BackendConfiguration
+from qiskit.backends.models.backendconfiguration import GateConfig
 from qiskit.qobj import Qobj, QobjItem, QobjConfig, QobjHeader, QobjInstruction
 from qiskit.qobj import QobjExperiment, QobjExperimentHeader
 from qiskit.backends.jobstatus import JobStatus
@@ -179,10 +180,31 @@ def new_fake_qobj():
         )]
     )
 
-
-class FakeBackend():
-    """Fakes qiskit.backends.basebackend.BaseBackend instances."""
+class FakeBackend(object):
+    """A fake backend.
+    """
 
     def name(self):
-        """Return the name of the backend."""
-        return 'test-backend'
+        """ name of fake backend"""
+        return 'qiskit_is_cool'
+
+    def configuration(self):
+        """Return a make up configuration for a fake device."""
+        qx5_cmap = [[1, 0], [1, 2], [2, 3], [3, 4], [3, 14], [5, 4], [6, 5],
+                    [6, 7], [6, 11], [7, 10], [8, 7], [9, 8], [9, 10], [11, 10],
+                    [12, 5], [12, 11], [12, 13], [13, 4], [13, 14], [15, 0],
+                    [15, 2], [15, 14]]
+        return BackendConfiguration(
+            backend_name='fake',
+            backend_version='0.0.0',
+            n_qubits=16,
+            basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
+            simulator=False,
+            local=True,
+            conditional=False,
+            open_pulse=False,
+            memory=False,
+            max_shots=65536,
+            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
+            coupling_map=qx5_cmap,
+        )
